+++ example/apps/functions/src/controllers/auth/on-auth-create.ts
@@ -1,0 +1,5 @@
+import { onAuthCreate } from 'nx-cloud-functions-deployer';
+
+export default onAuthCreate(({ uid }) => {
+	console.log('New user created: uid', uid, );
+});