+++ src/constants/function-types.ts
@@ -23,11 +23,19 @@ export const storageFunctions = [
 	'onObjectMetadataUpdated',
 ] as const;
 
+export const authFunctions = [
+	'onAuthCreate',
+	'onAuthDelete',
+	'beforeAuthCreate',
+	'beforeAuthSignIn',
+] as const;
+
 export const schedulerFunctions = ['onSchedule'] as const;
 
 export const httpsFunctions = ['onCall', 'onRequest'] as const;
 
 export const functions = [
+	...authFunctions,
 	...databaseFunctions,
 	...firestoreFunctions,
 	...storageFunctions,